--- old/x/trust/internal/types/key.go
@@ -1,12 +1,0 @@
-package types
-
-const (
-	// ModuleName is the name of the module
-	ModuleName = "trust"
-
-	// StoreKey to be used when creating the KVStore
-	StoreKey = ModuleName
-
-	// QueryAccountScores QueryAccountScores
-	QueryAccountScores = "account-scores"
-)
