--- old/x/trust/module.go
@@ -1,130 +1,0 @@
-package trust
-
-import (
-	"encoding/json"
-
-	"github.com/gorilla/mux"
-	"github.com/spf13/cobra"
-
-	"github.com/cosmos/cosmos-sdk/codec"
-	"github.com/cosmos/cosmos-sdk/types/module"
-	"github.com/lcnem/trust/x/trust/client/cli"
-	"github.com/lcnem/trust/x/trust/client/rest"
-
-	"github.com/cosmos/cosmos-sdk/client/context"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	abci "github.com/tendermint/tendermint/abci/types"
-)
-
-// type check to ensure the interface is properly implemented
-var (
-	_ module.AppModule      = AppModule{}
-	_ module.AppModuleBasic = AppModuleBasic{}
-)
-
-// AppModuleBasic is an app module Basics object
-type AppModuleBasic struct{}
-
-// Name returns module name
-func (AppModuleBasic) Name() string {
-	return ModuleName
-}
-
-// RegisterCodec returns RegisterCodec
-func (AppModuleBasic) RegisterCodec(cdc *codec.Codec) {
-	RegisterCodec(cdc)
-}
-
-// DefaultGenesis returns default genesis state
-func (AppModuleBasic) DefaultGenesis() json.RawMessage {
-	return ModuleCdc.MustMarshalJSON(DefaultGenesisState())
-}
-
-// ValidateGenesis checks the Genesis
-func (AppModuleBasic) ValidateGenesis(bz json.RawMessage) error {
-	var data GenesisState
-	err := ModuleCdc.UnmarshalJSON(bz, &data)
-	if err != nil {
-		return err
-	}
-	// Once json successfully marshalled, passes along to genesis.go
-	return ValidateGenesis(data)
-}
-
-// RegisterRESTRoutes returns rest routes
-func (AppModuleBasic) RegisterRESTRoutes(ctx context.CLIContext, rtr *mux.Router) {
-	rest.RegisterRoutes(ctx, rtr, StoreKey)
-}
-
-// GetQueryCmd returns the root query command of this module
-func (AppModuleBasic) GetQueryCmd(cdc *codec.Codec) *cobra.Command {
-	return cli.GetQueryCmd(StoreKey, cdc)
-}
-
-// GetTxCmd returns the root tx command of this module
-func (AppModuleBasic) GetTxCmd(cdc *codec.Codec) *cobra.Command {
-	return cli.GetTxCmd(StoreKey, cdc)
-}
-
-// AppModule struct
-type AppModule struct {
-	AppModuleBasic
-	keeper     Keeper
-}
-
-// NewAppModule creates a new AppModule Object
-func NewAppModule(k Keeper) AppModule {
-	return AppModule{
-		AppModuleBasic: AppModuleBasic{},
-		keeper:         k,
-	}
-}
-
-// Name returns module name
-func (AppModule) Name() string {
-	return ModuleName
-}
-
-// RegisterInvariants is empty
-func (am AppModule) RegisterInvariants(ir sdk.InvariantRegistry) {}
-
-// Route returns RouterKey
-func (am AppModule) Route() string {
-	return RouterKey
-}
-
-// NewHandler returns new Handler
-func (am AppModule) NewHandler() sdk.Handler {
-	return NewHandler(am.keeper)
-}
-
-// QuerierRoute returns module name
-func (am AppModule) QuerierRoute() string {
-	return ModuleName
-}
-
-// NewQuerierHandler returns new Querier
-func (am AppModule) NewQuerierHandler() sdk.Querier {
-	return NewQuerier(am.keeper)
-}
-
-// BeginBlock is a callback function
-func (am AppModule) BeginBlock(_ sdk.Context, _ abci.RequestBeginBlock) {}
-
-// EndBlock is a callback function
-func (am AppModule) EndBlock(sdk.Context, abci.RequestEndBlock) []abci.ValidatorUpdate {
-	return []abci.ValidatorUpdate{}
-}
-
-// InitGenesis inits genesis
-func (am AppModule) InitGenesis(ctx sdk.Context, data json.RawMessage) []abci.ValidatorUpdate {
-	var genesisState GenesisState
-	ModuleCdc.MustUnmarshalJSON(data, &genesisState)
-	return InitGenesis(ctx, am.keeper, genesisState)
-}
-
-// ExportGenesis exports genesis
-func (am AppModule) ExportGenesis(ctx sdk.Context) json.RawMessage {
-	gs := ExportGenesis(ctx, am.keeper)
-	return ModuleCdc.MustMarshalJSON(gs)
-}
