--- old/x/trust/spec/01-concepts.md
@@ -1,4 +1,0 @@
-# trust Concepts
-
-ページランクアルゴリズムによってアドレスをスコア付けする機能。
-トピックIDごとにスコアをつけられる。
