+++ x/trust/client/cli/query.go
@@ -28,7 +28,7 @@ func GetQueryCmd(queryRoute string, cdc *codec.Codec) *cobra.Command {
 
 	trustQueryCmd.AddCommand(
 		flags.GetCommands(
-	// TODO: Add query Cmds
+			getCmdAccountScores(queryRoute, cdc),
 		)...,
 	)
 
@@ -37,3 +37,26 @@ func GetQueryCmd(queryRoute string, cdc *codec.Codec) *cobra.Command {
 }
 
 // TODO: Add Query Commands
+
+func getCmdAccountScores(queryRoute string, cdc *codec.Codec) *cobra.Command {
+	return &cobra.Command{
+		Use:   "account-scores [address] [topic_ids[,]]",
+		Short: "get account scores",
+		Args:  cobra.ExactArgs(2),
+		RunE: func(cmd *cobra.Command, args []string) error {
+			cliCtx := context.NewCLIContext().WithCodec(cdc)
+
+			bz, _ := cdc.MarshalJSON(types.QueryAccountScoresParam{Address: args[0], TopicIDs: args[1]})
+
+			res, _, err := cliCtx.QueryWithData(fmt.Sprintf("custom/%s/%s", queryRoute, types.QueryAccountScores), bz)
+			if err != nil {
+				fmt.Printf(err.Error())
+				return nil
+			}
+			var scores map[string]float64
+			json.Unmarshal(res, &scores)
+
+			return cliCtx.PrintOutput(scores)
+		},
+	}
+}
