--- x/trust/internal/keeper/params.go
@@ -1,22 +1,0 @@
-package keeper
-/*
-// TODO: Define if your module needs Parameters, if not this can be deleted
-
-import (
-	"time"
-
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/lcnem/trust/x/trust/internal/types"
-)
-
-// GetParams returns the total set of trust parameters.
-func (k Keeper) GetParams(ctx sdk.Context) (params types.Params) {
-	k.paramspace.GetParamSet(ctx, &params)
-	return params
-}
-
-// SetParams sets the trust parameters to the param space.
-func (k Keeper) SetParams(ctx sdk.Context, params types.Params) {
-	k.paramspace.SetParamSet(ctx, &params)
-}
-*/
