+++ x/trust/internal/types/errors.go
@@ -8,6 +8,7 @@ import (
 
 // TODO: Fill out some custom errors for the module
 // You can see how they are constructed below:
-// var (
-//	ErrInvalid = sdkerrors.Register(ModuleName, 1, "custom error message")
-// )
+var (
+	ErrInvalidTopicID = sdkerrors.Register(ModuleName, 1, "topic id must start with [a-z] and available characters are [a-z],[0-9],-")
+	ErrInvalidWeight = sdkerrors.Register(ModuleName, 2, "weight must be positive")
+)
