--- x/trust/internal/types/expected_keepers.go
@@ -1,24 +1,0 @@
-package types
-
-import (
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/x/params"
-)
-
-// ParamSubspace defines the expected Subspace interfacace
-type ParamSubspace interface {
-	WithKeyTable(table params.KeyTable) params.Subspace
-	Get(ctx sdk.Context, key []byte, ptr interface{})
-	GetParamSet(ctx sdk.Context, ps params.ParamSet)
-	SetParamSet(ctx sdk.Context, ps params.ParamSet)
-}
-
-/*
-When a module wishes to interact with an otehr module it is good practice to define what it will use
-as an interface so the module can not use things that are not permitted.
-TODO: Create interfaces of what you expect the other keepers to have to be able to use this module.
-type BankKeeper interface {
-	SubtractCoins(ctx sdk.Context, addr sdk.AccAddress, amt sdk.Coins) (sdk.Coins, error)
-	SendCoins(ctx sdk.Context, fromAddr sdk.AccAddress, toAddr sdk.AccAddress, amt sdk.Coins) error
-}
-*/
